One class per PHP file
