+++ src/AdjustmentOfferingResult.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+class AdjustmentOfferingResult {
+    public string $vendorId;
+    public string $adjustmentId;
+    public string $vendorSaleReference;
+    public bool $sold;
+}
