+++ src/AdjustmentRequest.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+class AdjustmentRequest {
+    public string $vendorId;
+    public string $offeringId;
+    public OfferingRequest $offeringRequest;
+}
