+++ src/ApiClientConfiguration.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+class ApiClientConfiguration {
+    public string $apiKey;
+    public string $vendorId;
+    public string $certificatePath;
+    public string $environment;
+}
