+++ src/CurlResponse.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+class CurlResponse {
+    public int $statusCode;
+    public ?string $responseBody;
+}
