+++ src/IAuthTokenGenerator.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+interface IAuthTokenGenerator {
+    public function generateToken(string $vendorId, string $apiKey): string;
+}
