+++ src/InsureHubException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+use Exception;
+
+abstract class InsureHubException extends Exception {
+    abstract public function errorMessage();
+}
