+++ src/MatrixRequest.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+class MatrixRequest {
+    public string $productCode;
+    public string $currencyCode;
+}
