+++ src/Offering.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+class Offering {
+    public function __construct() {
+        $this->productOfferings         = [];
+        $this->productsOutsideOfPricing = [];
+    }
+
+    public string $id;
+    /**
+     * @var ProductOffering[]
+     */
+    public array $productOfferings;
+    /**
+     * @var Product[]
+     */
+    public ?array $productsOutsideOfPricing;
+}
