+++ src/PriceBand.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+class PriceBand {
+    public string $id;
+    public float $lowerBound;
+    public float $upperBound;
+    public float $premium;
+    public float $commission;
+}
