+++ src/PriceBandRequest.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+class PriceBandRequest extends MatrixRequest {
+    public float $price;
+}
