+++ src/ProductOfferingWording.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace BookingProtect\InsuranceHub\Client;
+
+class ProductOfferingWording {
+    public string $categoryCode;
+    public string $description;
+    public string $languageCode;
+    public string $salesProcessCode;
+    public string $salesProcessMessage;
+    public string $salesMessage;
+    public ?string $advertisement;
+    public ?string $advertisementUrl;
+    public string $logoUrl;
+    public string $termsAndConditions;
+}
